Example of Capturing ContactChangeEvents
This example shows how to inherit from accc_CDCEventBase to handle ContactChangeEvents along with unit tests to ensure 100% code coverage. The wiring in the sf-trigger-handler is present in the current ACCC but commented out. The only thing required are 
(1) uncomment out the accc_cdcContactHandler.trigger and 
(2) uncomment out the Trigger Mediator in accc_ApexPlatformEventDispatcher
(3) Once the sf-trigger-handler is installed, add references for ContactChangeEvent (to call ContactChangeEventHandler) for both test, debug and production environment.
(4) Go into setup and enable CDC for Contact